pyproject.toml
[tool.poetry]
name = "django-learning"
version = "0.1.0"
description = ""
authors = ["Rodrigo Diaz <rodrigodiaz@frba.utn.edu.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
